requirements.txt
Flask>=3.0,<4.0
Pillow>=10.0,<11.0
opencv-python-headless>=4.9,<5.0
